+++ cli/state.ts
@@ -1,6 +1,6 @@
-import { Denomander, exists, format } from "../deps.ts";
+import { CliffySelect, exists, format } from "../deps.ts";
 import { isUrl, parsePath } from "./utils.ts";
-import type { NessieConfig } from "../types.ts";
+import type { CommandOptions, NessieConfig } from "../types.ts";
 import {
   DEFAULT_CONFIG_FILE,
   MAX_FILE_NAME_LENGTH,
@@ -17,9 +17,9 @@ export class State {
   private config?: NessieConfig;
   client?: NessieConfig["client"];
 
-  constructor(args: Denomander) {
-    this.enableDebug = args.debug;
-    this.configFile = parsePath(args.config || DEFAULT_CONFIG_FILE);
+  constructor(options: CommandOptions) {
+    this.enableDebug = !!options.debug;
+    this.configFile = parsePath(options.config || DEFAULT_CONFIG_FILE);
 
     this.logger([this.enableDebug, this.configFile], "State");
   }
@@ -77,17 +77,21 @@ export class State {
       this.logger(fileName, "Migration file name");
     }
 
-    await Deno.mkdir(this.client!.migrationFolder, { recursive: true });
+    const selectedFolder = await this._folderPrompt(
+      this.client!.migrationFolders,
+    );
+
+    await Deno.mkdir(selectedFolder, { recursive: true });
 
     const responseFile = await fetch(URL_TEMPLATE_BASE + "/migration.ts");
 
     await Deno.writeTextFile(
-      `${this.client!.migrationFolder}/${fileName}`,
+      `${selectedFolder}/${fileName}`,
       await responseFile.text(),
     );
 
     console.info(
-      `Created migration ${fileName} at ${this.client!.migrationFolder}`,
+      `Created migration ${fileName} at ${selectedFolder}`,
     );
   }
 
@@ -100,17 +104,19 @@ export class State {
 
     this.logger(fileName, "Seed file name");
 
-    await Deno.mkdir(this.client!.seedFolder, { recursive: true });
+    const selectedFolder = await this._folderPrompt(this.client!.seedFolders);
+
+    await Deno.mkdir(selectedFolder, { recursive: true });
 
     const responseFile = await fetch(URL_TEMPLATE_BASE + "seed.ts");
 
     await Deno.writeTextFile(
-      `${this.client!.seedFolder}/${fileName}`,
+      `${selectedFolder}/${fileName}`,
       await responseFile.text(),
     );
 
     console.info(
-      `Created seed ${fileName} at ${this.client!.seedFolder}`,
+      `Created seed ${fileName} at ${selectedFolder}`,
     );
   }
 
@@ -125,5 +131,26 @@ export class State {
     } catch {
       console.error("Error at: " + title);
     }
+  }
+
+  private async _folderPrompt(folders: string[]) {
+    let promptSelection = 0;
+
+    if (folders.length > 1) {
+      const promptResult = await CliffySelect.prompt({
+        message:
+          `You have multiple folder sources, where do you want to create the new file?`,
+        options: folders.map((folder, i) => ({
+          value: i.toString(),
+          name: folder,
+        })),
+      });
+
+      promptSelection = parseInt(promptResult);
+    }
+
+    this.logger(promptSelection, "Prompt input final");
+
+    return folders[promptSelection];
   }
 }