+++ tests/cli-experimental/config/sqlite3.config.ts
@@ -3,8 +3,14 @@ import { ClientSQLite } from "../../../mod.ts";
 export default {
   client: new ClientSQLite(
     {
-      migrationFolder: "./tests/cli-experimental/sqlite3",
-      seedFolder: "./tests/cli-experimental/sqlite3",
+      migrationFolders: [
+        "./tests/cli-experimental/sqlite31",
+        "./tests/cli-experimental/sqlite32",
+      ],
+      seedFolders: [
+        "./tests/cli-experimental/sqlite31",
+        "./tests/cli-experimental/sqlite32",
+      ],
     },
     "./tests/data/sqlite.db",
   ),